Rename folder and solution
