+++ src/ParquetViewer/AppSettings.cs
