+++ src/ParquetViewer/Controls/DelayedOnChangedTextBox.cs
