+++ src/ParquetViewer/Helpers/CustomScriptBasedSchemaAdapter.cs
