+++ src/ParquetViewer/Helpers/ExtensionMethods.cs
