+++ src/ParquetViewer/Helpers/ParquetMetadataAnalyzers.cs
