+++ src/ParquetViewer/Helpers/UtilityMethods.cs
