+++ src/ParquetViewer/MainForm.EventHandlers.cs
