+++ src/ParquetViewer/MainForm.Helpers.cs
