+++ src/ParquetViewer/MetadataViewer.cs
