+++ src/ParquetViewer/Properties/AssemblyInfo.cs
@@ -1,0 +1,36 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+using System.Runtime.Versioning;
+
+// General Information about an assembly is controlled through the following 
+// set of attributes. Change these attribute values to modify the information
+// associated with an assembly.
+[assembly: AssemblyTitle("Parquet Viewer")]
+[assembly: AssemblyDescription("Simple windows desktop application for viewing Apache Parquet files\r\nhttps://github.com/mukunku/ParquetViewer\r\n\r\nSpecial thanks to: https://github.com/aloneguid/parquet-dotnet")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("Author: Mukunku")]
+[assembly: AssemblyProduct("Parquet Viewer")]
+[assembly: AssemblyCopyright("GNU General Public License v3.0")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+[assembly: SupportedOSPlatform("windows")]
+
+// Setting ComVisible to false makes the types in this assembly not visible 
+// to COM components.  If you need to access a type in this assembly from 
+// COM, set the ComVisible attribute to true on that type.
+[assembly: ComVisible(false)]
+
+// The following GUID is for the ID of the typelib if this project is exposed to COM
+[assembly: Guid("134e9435-2fb2-4297-ac49-d44477a427bc")]
+
+// Version information for an assembly consists of the following four values:
+//
+//      Major Version
+//      Minor Version 
+//      Build Number
+//      Revision
+//
+// You can specify all the values or you can default the Build and Revision Numbers 
+// by using the '*' as shown below:
+// [assembly: AssemblyVersion("1.0.*")]
+[assembly: AssemblyVersion("2.6.0.0")]
