+++ src/ParquetViewer/Resources/coffee.gif
