+++ src/ParquetViewer/Resources/exclamation_icon.png
