+++ src/ParquetViewer/Resources/hourglass.gif
