+++ src/ParquetViewer/Resources/parquet_icon.ico
