+++ src/ParquetViewer/Resources/plus_minus_icon.ico
